migrate to 1d generic positioner
