+++ src/dodal/beamlines/i18.py
@@ -13,8 +13,6 @@ from dodal.common.visit import (
     StaticVisitPathProvider,
 )
 from dodal.devices.common_dcm import BaseDCM, PitchAndRollCrystal, RollCrystal
-from dodal.devices.i18.d7positionerA import D7PositionerA
-from dodal.devices.i18.d7positionerB import D7PositionerB
 from dodal.devices.i18.diode import Diode
 from dodal.devices.i18.KBMirror import KBMirror
 from dodal.devices.i18.table import Table
@@ -117,16 +115,6 @@ def hfm() -> KBMirror:
 @device_factory()
 def d7_diode() -> Diode:
     return Diode(f"{PREFIX.beamline_prefix}-DI-PHDGN-07:")
-
-
-@device_factory()
-def d7_positioner_b() -> D7PositionerB:
-    return D7PositionerB(f"{PREFIX.beamline_prefix}-DI-PHDGN-07:B:MP")
-
-
-@device_factory()
-def d7_filter_positioner() -> D7PositionerA:
-    return D7PositionerA(f"{PREFIX.beamline_prefix}-DI-PHDGN-07:A:MP")
 
 
 @device_factory()
